Consolidate query tabs
This moves all query tabs into the same panel. The reason it's a large
commit is because it refactors the TabsPanel component to accept the
selected tab as a prop instead of managing it internally. This was
necessary in order to programmatically switch to the query editor tab
after selecting an example.

This changes the way that tabs are defined to make them keyed rather
than a list with an `id` field, which simplifies defining what tabs are
rendered in panels. (Defining them, rather than relying on list
filtering).

It also moves tab definitions from the routes files, into the Panel file
where they reside.
