+++ src/routes/term.$termId.tsx
@@ -1,11 +1,9 @@
 import { styled } from "@mui/joy";
-import { createFileRoute, getRouteApi, Link } from "@tanstack/react-router";
+import { createFileRoute } from "@tanstack/react-router";
 import { useWindowSize } from "@uidotdev/usehooks";
 import { TermPagePanels } from "../ui/Panels/layout/TermPagePanels";
 import { useEffect, useState } from "react";
 import { QueryEngine } from "@comunica/query-sparql";
-import { RDFTable } from "../ui/RDFTable/RDFTable";
-import { useComunicaQuery } from "../hooks/useComunicaQuery";
 import dedent from "dedent";
 
 const engine = new QueryEngine();
@@ -88,119 +86,8 @@ function RouteComponent() {
         </a>
       </header>
 
-      <TermPagePanels
-        tabs={[
-          {
-            id: "as-subject",
-            label: "As Subject",
-            color: "var(--p-orange-400)",
-            jsx: (
-              <TermPanel
-                querySparql={dedent`
-                  SELECT ?predicate ?object
-                  WHERE {
-                    <${termId}> ?predicate ?object
-                    FILTER(!isLiteral(?object))
-                  }
-                  LIMIT 50
-                `}
-              />
-            ),
-          },
-          {
-            id: "as-object",
-            label: "As Object",
-            color: "var(--p-cyan-400)",
-            jsx: (
-              <TermPanel
-                querySparql={dedent`
-                  SELECT ?subject ?predicate
-                  WHERE {
-                    ?subject ?predicate <${termId}>
-                    FILTER(!isLiteral(?subject))
-                  }
-                  LIMIT 50
-                `}
-              />
-            ),
-          },
-          {
-            id: "attributes",
-            label: "Attributes",
-            color: "var(--p-indigo-300)",
-            jsx: (
-              <TermPanel
-                querySparql={dedent`
-                  SELECT ?property ?value
-                  WHERE {
-                    <${termId}> ?property ?value
-                    FILTER(isLiteral(?value))
-                  }
-                  LIMIT 50
-                `}
-              />
-            ),
-          },
-          {
-            id: "as-predicate",
-            label: "As Predicate",
-            color: "var(--p-pink-400)",
-            jsx: (
-              <TermPanel
-                querySparql={dedent`
-                  SELECT ?subject ?object
-                  WHERE {
-                    ?subject <${termId}> ?object
-                  }
-                  LIMIT 50
-                `}
-              />
-            ),
-          },
-        ]}
-      />
+      <TermPagePanels termId={termId} />
     </Wrapper>
-  );
-}
-
-const rootRouteApi = getRouteApi("__root__");
-
-interface TermPanelProps {
-  querySparql: string;
-}
-function TermPanel({ querySparql }: TermPanelProps) {
-  const { sources } = rootRouteApi.useLoaderData();
-
-  const query = useComunicaQuery({
-    runOnMount: true,
-    query: querySparql,
-    sources: sources.filter((s) => s.shortname === "federation"),
-  });
-
-  if (query.isRunning) return <div>Loading...</div>;
-
-  return (
-    <PanelWrapper>
-      <TableHeader>
-        <Link
-          to="/"
-          search={{
-            query: querySparql.replace("LIMIT 50", "LIMIT 1000"),
-            sources: ["federation"],
-          }}
-        >
-          First 50 results, click to view full query.
-        </Link>
-      </TableHeader>
-      <TableWrapper>
-        <RDFTable
-          columns={query.columns}
-          rows={query.results}
-          wrapText={true}
-          resolveLabels={true}
-        />
-      </TableWrapper>
-    </PanelWrapper>
   );
 }
 
@@ -238,23 +125,4 @@ const Heading = styled("h1")`
     left: 0px;
     right: 0px;
   }
-`;
-
-const PanelWrapper = styled("div")`
-  --parent-padding: 0.75rem;
-  height: calc(100% + 2 * var(--parent-padding));
-  margin: calc(-1 * var(--parent-padding));
-  display: flex;
-  flex-direction: column;
-`;
-
-const TableWrapper = styled("div")`
-  flex: 1;
-  min-height: 0px;
-`;
-
-const TableHeader = styled("header")`
-  padding: 0.75rem 1rem 0rem 1rem;
-  gap: 0.5rem;
-  background-color: var(--p-slate-50);
 `;
